feat: initial commit of tic-tac-toe
Prompt:

We are aiming to build tic-tac-toe. This is the starter template (vite)
in TS. We will remove this code and make Tic Tac Toe.

In the @Game.ts file, create tic-tac-toe. The game should start with
the player being random X or O. They should follow the game rules until
the game is over. The winner should be declared. They are playing
against a basic AI.

After the game they can click a button to restart it. When they do so,
they swap from X to O or vice versa and play again.