feat: tic tac toe (#1)
* feat: initial commit of tic-tac-toe

Prompt:

We are aiming to build tic-tac-toe. This is the starter template (vite)
in TS. We will remove this code and make Tic Tac Toe.

In the @Game.ts file, create tic-tac-toe. The game should start with
the player being random X or O. They should follow the game rules until
the game is over. The winner should be declared. They are playing
against a basic AI.

After the game they can click a button to restart it. When they do so,
they swap from X to O or vice versa and play again.

* test: add tests to game logic

Prompt:

Now we need to add some tests (vitest)

We want to test the game logic; not the drawing logic, but the game
logic. We want to test the end conditions as well as the AI logic.

Do that, and refactor the game scene as necessary to make it easy to
test.

* chore: cleanup